fix(test-kit): make fixture install race-free and self-sufficient
The Tech Lead review found that buildSite() required the fixture's
node_modules to exist beforehand, but neither the test:server npm script
nor the lint-and-server CI job installed them. Cold-checkout npm test
exited 1.

Fix has three parts, all reviewer-recommended:

1. Extract ensureFixtureInstalled() into its own helper module so both
   buildSite() and generateSite() can share it without duplicating code
   or creating a circular import.
2. Call ensureFixtureInstalled() inside buildSite() so direct callers
   (one-off scripts, debug sessions) work from a fresh clone.
3. Add tests/helpers/install-fixture.js as a CLI shim and prepend it to
   the test:server npm script. node --test forks one subprocess per
   test file; without a single-process precondition step they would
   race each other on the very first run and corrupt the install with
   EEXIST on the hexo-blog-encrypt symlink. test:e2e is unaffected
   because Playwright's globalSetup is single-process.
4. Add explicit 'Install fixture site dependencies' step to the
   lint-and-server CI job. Mirrors the e2e job's pattern, surfaces
   install time in CI logs, and keeps coverage timing clean.

Verified: with tests/fixtures/hexo-site/node_modules removed,
'npm run test:server' exits 0 with 23/23 pass and 100% statement
coverage on index.js.
