+++ .github/workflows/test.yml
@@ -16,6 +16,9 @@ jobs:
       - name: Install root dependencies
         run: npm install --no-audit --no-fund
 
+      - name: Install fixture site dependencies
+        run: cd tests/fixtures/hexo-site && npm install --no-audit --no-fund && cd ../../..
+
       - name: Lint
         run: npm run lint
